Use canlib util to generate timing parameters
This commit revs the version of canlib so that we can use the
can_generate_timing_params function instead of writing out the timing
manually.

Note, this change required changing nbproject/configurations.xml.
Because Alex is on an older version of MPLAB, this would break her
build. So this commit does not add that. You need to add the
timing_util.{c,h} files to your project config in order for this to run
properly.