Set linear actuator DAC value
We swapped the open DAC with a resistive voltage divider so that it
would work even before the vent microcontroller woke up. The vent valve
hasn't been hydrostatic tested with the newly set DAC value, but it
seems to be pretty tightly closed, so I'm happy with it. Hydrostatic
test incoming either tomorrow or Sunday.

Because the resistive divider doesn't need initialization, the code to
initialize the open DAC is removed. The LOW_DAC_VAL macro is removed,
since it wasn't being used.